refactor: made `whittaker_base` a private utility
