+++ chemotools/utils/_whittaker_base/main.py
@@ -17,10 +17,10 @@ import numpy as np
 from chemotools._runtime import PENTAPY_AVAILABLE
 from chemotools.utils import models
 from chemotools.utils.banded_linalg import LAndUBandCounts
-from chemotools.utils.whittaker_base import auto_lambda as auto
-from chemotools.utils.whittaker_base import initialisation as init
-from chemotools.utils.whittaker_base import solvers
-from chemotools.utils.whittaker_base.misc import get_weight_generator
+from chemotools.utils._whittaker_base import auto_lambda as auto
+from chemotools.utils._whittaker_base import initialisation as init
+from chemotools.utils._whittaker_base import solvers
+from chemotools.utils._whittaker_base.misc import get_weight_generator
 
 ### Class Implementation ###
